+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - **breaking:** alias: display an error with `/alias add` or `/alias addcompletion` when the alias already exists, add options `addreplace` and `addreplacecompletion` in command `/alias` ([#2095](https://github.com/weechat/weechat/issues/2095))
+- **breaking:** core: add option `addreplace` in command `/bar` ([#2095](https://github.com/weechat/weechat/issues/2095))
 - relay: enable websocket extension "permessage-deflate" with "api" relay only ([#1549](https://github.com/weechat/weechat/issues/1549))
 
 ### Added